requirements.txt
python-dotenv>=1.0.1
certifi>=2024.0.0
SQLAlchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.3
aiogram>=3.4.0
beautifulsoup4>=4.12.3
aiohttp>=3.9.3

asyncpg>=0.29.0

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.1.0

alembic>=1.13.0
firebase-admin>=6.4.0
aiosmtplib>=3.0.0
slowapi>=0.1.9

pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=4.1
aioresponses>=0.7.6